Fix another Windows 64-bit build warning

+++ suite/tests/client-interface/drbbdup-emul-test.c
@@ -41,7 +41,7 @@ main(int argc, char **argv)
     char buf2[1024];
 #    ifdef WINDOWS
 #        ifdef X64
-    __movsq((unsigned long *)buf2, (unsigned long *)buf1, 0);
+    __movsq((unsigned long long *)buf2, (unsigned long long *)buf1, 0);
 #        else
     __movsd((unsigned long *)buf2, (unsigned long *)buf1, 0);
 #        endif
